requirements.txt
ultralytics
transformers
torch
torchvision
requests
opencv-python
pillow
numpy
python-dotenv
onnx
onnxslim
onnxruntime